#10076: Merge rad2deg_bw to TTNN
