+++ docs/source/ttnn/ttnn/ttnn/rad2deg_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.rad2deg_bw:
+
+ttnn.rad2deg_bw
+################
+
+.. autofunction:: ttnn.rad2deg_bw
