+++ tt_eager/tt_dnn/op_library/backward/backward_ops.cpp
@@ -361,41 +361,6 @@ std::vector<Tensor> bias_gelu_unary_bw(
         grad, input, bias, approximate, output_mem_config);
 }
 
-std::vector<Tensor> _hardshrink_bw(
-    const Tensor& grad, const Tensor& input_tensor, float lambd, const MemoryConfig& output_mem_config) {
-    std::vector<Tensor> grad_tensor;
-    Tensor hardshrink_result = hardshrink(input_tensor, lambd, output_mem_config);
-    Tensor result = where(ttnn::eqz(hardshrink_result, output_mem_config), 0.0f, grad, output_mem_config);
-    grad_tensor.emplace_back(result);
-    return grad_tensor;
-}
-std::vector<Tensor> hardshrink_bw(
-    const Tensor& grad, const Tensor& input, float lambd, const MemoryConfig& output_mem_config) {
-    return operation::decorate_as_composite(__func__, _hardshrink_bw)(grad, input, lambd, output_mem_config);
-}
-
-// softshrink
-//  result: torch.where(self < -lambd, grad, torch.where(self > lambd, grad, torch.tensor(0.0)))
-std::vector<Tensor> _softshrink_bw(
-    const Tensor& grad, const Tensor& input_tensor, float lambd, const MemoryConfig& output_mem_config) {
-    std::vector<Tensor> grad_tensor;
-    Tensor result = where(
-        ttnn::logical_or(
-            ttnn::lt(input_tensor, full_like(input_tensor, -lambd, output_mem_config), std::nullopt, output_mem_config),
-            ttnn::gt(input_tensor, full_like(input_tensor, lambd, output_mem_config), std::nullopt, output_mem_config),
-            std::nullopt,
-            output_mem_config),
-        grad,
-        zeros_like(grad, output_mem_config),
-        output_mem_config);
-    grad_tensor.emplace_back(result);
-    return grad_tensor;
-}
-std::vector<Tensor> softshrink_bw(
-    const Tensor& grad, const Tensor& input, float lambd, const MemoryConfig& output_mem_config) {
-    return operation::decorate_as_composite(__func__, _softshrink_bw)(grad, input, lambd, output_mem_config);
-}
-
 // Hardswish
 // result: torch.where(input < -3,0.0,torch.where(input <= 3, grad * ((input / 3) + 0.5), grad),)
 std::vector<Tensor> _hardswish_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
@@ -610,39 +575,6 @@ std::vector<Tensor> _cosh_bw(const Tensor& grad, const Tensor& input, const Memo
 }
 std::vector<Tensor> cosh_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
     return operation::decorate_as_composite(__func__, _cosh_bw)(grad, input, output_mem_config);
-}
-
-// Leaky_Relu
-// result: torch.where(self > 0, grad_output, grad_output * negative_slope)
-std::vector<Tensor> _leaky_relu_bw(
-    const Tensor& grad, const Tensor& input, float negative_slope, const MemoryConfig& output_mem_config) {
-    std::vector<Tensor> grad_tensor;
-    Tensor grad_result = where(
-        ttnn::gtz(input, output_mem_config), grad, ttnn::multiply(grad, negative_slope, std::nullopt, output_mem_config), output_mem_config);
-    grad_tensor.emplace_back(grad_result);
-    return grad_tensor;
-}
-std::vector<Tensor> leaky_relu_bw(
-    const Tensor& grad, const Tensor& input, float negative_slope, const MemoryConfig& output_mem_config) {
-    return operation::decorate_as_composite(__func__, _leaky_relu_bw)(grad, input, negative_slope, output_mem_config);
-}
-
-// ELU
-// result : grad * (torch.where(input >= 0, 1, alpha * torch.exp(input)))
-std::vector<Tensor> _elu_bw(
-    const Tensor& grad, const Tensor& input, float alpha, const MemoryConfig& output_mem_config) {
-    std::vector<Tensor> grad_tensor;
-    Tensor grad_result = where(
-        ttnn::gez(input, output_mem_config),
-        grad,
-        ttnn::multiply(grad, ttnn::multiply(ttnn::exp(input, false, output_mem_config), alpha, std::nullopt, output_mem_config), std::nullopt, output_mem_config),
-        output_mem_config);
-    grad_tensor.emplace_back(grad_result);
-    return grad_tensor;
-}
-std::vector<Tensor> elu_bw(
-    const Tensor& grad, const Tensor& input, float alpha, const MemoryConfig& output_mem_config) {
-    return operation::decorate_as_composite(__func__, _elu_bw)(grad, input, alpha, output_mem_config);
 }
 
 // Hardtanh
@@ -869,16 +801,6 @@ std::vector<Tensor> deg2rad_bw(const Tensor& grad, const Tensor& input, const Me
     return operation::decorate_as_composite(__func__, _deg2rad_bw)(grad, input, output_mem_config);
 }
 
-std::vector<Tensor> _rad2deg_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    std::vector<Tensor> grad_tensor;
-    float M_180_PI = 180 / M_PI;
-    Tensor grad_result = ttnn::multiply(grad, M_180_PI, std::nullopt, output_mem_config);
-    grad_tensor.emplace_back(grad_result);
-    return grad_tensor;
-}
-std::vector<Tensor> rad2deg_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    return operation::decorate_as_composite(__func__, _rad2deg_bw)(grad, input, output_mem_config);
-}
 
 std::vector<Tensor> _reciprocal_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
     std::vector<Tensor> grad_tensor;
